pyproject.toml
[project]
name = "artefactual"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hicham Randrianarivo", email = "hicham.randrianarivo@gmail.com" },
]
requires-python = ">=3.10,<3.12"
dependencies = [
    "absl-py>=2.1.0",
    "beartype>=0.19.0",
    "etils[eapp,edc,enp,epath,epy,etqdm,etree]>=1.11.0",
    "keras-hub>=0.18.1",
    "keras>3",
    "scikit-learn>=1.6.0",
    "tensorflow-datasets>=4.9.7",
    "jax[cuda12]; platform_system != 'Darwin'",
    "jax; platform_system == 'Darwin'",
    "clu>=0.0.12",
    "tensorflow-cpu; platform_system != 'Darwin'",
    "tensorflow; platform_system == 'Darwin'",
]

[project.scripts]
artefactual = "artefactual:main"

[project.optional-dependencies]
hub = ["huggingface-hub>=0.27.0", "keras-hub>=0.18.1", "safetensors>=0.4.5"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "datasets>=3.2.0",
    "einops>=0.8.0",
    "faiss-cpu>=1.9.0.post1",
    "fastembed>=0.4.2 ; sys_platform == 'darwin'",
    "grain>=0.2.3",
    "ipython>=8.31.0",
    "jinja2>=3.1.5",
    "jupytext>=1.16.6",
    "lm-eval>=0.4.7 ; sys_platform == 'darwin'",
    "lm-eval>=0.4.7",
    "plum-dispatch>=2.5.7",
    "scikeras>=0.13.0",
    "toolz>=1.0.0",
    "wandb>=0.19.4",
    "lm-eval[vllm]>=0.4.7 ; sys_platform != 'darwin'",
    "mlcroissant>=1.0.12",
    "polars>=1.21.0",
    "orjson>=3.10.15",
    "tokenizers>=0.21.0",
    "transformers>=4.48.1",
    "outlines>=0.1.11",
    "pudb>=2024.1.3",
]

[tool.basedpyright]
typeCheckingMode = "recommended"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnannotatedClassAttribute = false
reportUninitializedInstanceVariable = false
reportOptionalSubscript = false
reportExplicitAny = false
reportAssignmentType = "warning"
enableExperimentalFeatures = true

[tool.ruff]
preview = true
target-version = "py310"
line-length = 120

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80

[tool.ruff.lint]
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT",
    "I",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "Q",
    "RUF",
    "S",
    "T",
    "TID",
    "UP",
    "W",
    "YTT",
]

ignore = ["ISC001"]

[tool.ruff.lint.per-file-ignores]
"**/scripts/*" = ["INP001", "T201", "PLR0917", "PLR0913", "RUF009", "FBT003"]
"**/tests/**/*" = ["PLC1901", "PLR2004", "PLR6301", "S", "TID252"]
"docs/notebooks/**/*" = ["INP001", "T201", "A004", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["argimi"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.jupytext]
cell_markers = '"""'
